requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.0.3
oauthlib==3.1.0
pycryptodome==3.9.0
requests==2.22.0
selenium==3.141.0